# swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Event Management Platform API
  description: Backend engine for a multi-event management platform
  version: 1.0.0
servers:
  - url: https://jan-two.vercel.app/api
    description: Production server
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /admin/events:
    get:
      summary: List all events managed by the admin
      tags:
        - Events (Admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of events
    post:
      summary: Create a new event
      tags:
        - Events (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, startDate, endDate, maxTeamSize]
              properties:
                title:
                  type: string
                description:
                  type: string
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
                venue:
                  type: string
                maxTeamSize:
                  type: integer
                registrationStart:
                  type: string
                  format: date-time
                registrationEnd:
                  type: string
                  format: date-time
                schedules:
                  type: array
                  items:
                    type: object
                    properties:
                      title: { type: string }
                      startTime: { type: string, format: date-time }
                      endTime: { type: string, format: date-time }
                      venue: { type: string }
                customFields:
                  type: array
                  items:
                    type: object
                    properties:
                      label: { type: string }
                      type: { type: string, enum: [TEXT, NUMBER, BOOLEAN] }
                      required: { type: boolean }
      responses:
        '201':
          description: Event created
...
  /events:
    get:
      summary: List all upcoming events
      tags:
        - Events (Public/Participant)
      parameters:
        - name: domain
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of upcoming events
...